requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-apscheduler==1.13.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
python-dateutil==2.8.2
pytest==7.4.3
flask-cors==4.0.0
